Shows Data for the Current Month
MySQL tables for rainlogger and levelogger were changed to better reflect the Access file coming in through the STS. The text files were updated to reflect the changes for the tables and data.

Updated the MySQL statement to gather data from the current month. For the rainlogger, the SQL statement sums up the amount of rainfall for each day of the month. For the levelogger, the SQL statement takes the average level of the water for each day of the month.

Commented out calendar and dropdown for now so that the next group can work on it, This is due to the issue of making an SQL connection in a front-end JS file. Also updated some variables and the filepaths.