Fixed graph to better fit graph cards
In the HTML, there is a "max-width" style set to the canvas element. That was changed to let the graph take up more of the graph card that was made bigger.

Switched the positions of the rainlogger and levelogger graphs.

Also changed variable names in graphs.js that ended in "1" or "2" to end in "Rainlogger" and "Levelogger".